+++ common/mp.cpp
@@ -107,7 +107,7 @@
  * HISTORY:                                                                                    *
  *   07/01/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-static int _Byte_Precision(unsigned long value)
+static int _Byte_Precision(uint32_t value)
 {
     int byte_count;
     for (byte_count = sizeof(value); byte_count; byte_count--) {
@@ -137,7 +137,7 @@ static int _Byte_Precision(unsigned long value)
  * HISTORY:                                                                                    *
  *   07/01/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-int XMP_DER_Length_Encode(unsigned long length, unsigned char* output)
+int XMP_DER_Length_Encode(uint32_t length, unsigned char* output)
 {
     assert(output != NULL);
 
@@ -1070,16 +1070,16 @@ bool XMP_Add_Int(digit* result, const digit* left_number, digit right_number, bo
  *=============================================================================================*/
 bool XMP_Sub(digit* result, const digit* left_number, const digit* right_number, bool borrow, int precision)
 {
-    const unsigned short* left_number_ptr = (const unsigned short*)left_number;
-    const unsigned short* right_number_ptr = (const unsigned short*)right_number;
-    unsigned short* result_ptr = (unsigned short*)result;
+    const uint16_t* left_number_ptr = (const uint16_t*)left_number;
+    const uint16_t* right_number_ptr = (const uint16_t*)right_number;
+    uint16_t* result_ptr = (uint16_t*)result;
 
     precision *= 2;
     while (precision--) {
         digit x = (digit)*left_number_ptr - (digit)*right_number_ptr - (digit)borrow;
         right_number_ptr++;
         left_number_ptr++;
-        *result_ptr++ = (unsigned short)x;
+        *result_ptr++ = (uint16_t)x;
         borrow = (((1L << 16) & x) != 0L);
     }
     return (borrow);
@@ -1110,16 +1110,16 @@ bool XMP_Sub(digit* result, const digit* left_number, const digit* right_number,
  * HISTORY:                                                                                    *
  *   07/01/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-bool XMP_Sub_Int(digit* result, const digit* left_number, unsigned short right_number, bool borrow, int precision)
+bool XMP_Sub_Int(digit* result, const digit* left_number, uint16_t right_number, bool borrow, int precision)
 {
-    const unsigned short* left_number_ptr = (const unsigned short*)left_number;
-    unsigned short* result_ptr = (unsigned short*)result;
+    const uint16_t* left_number_ptr = (const uint16_t*)left_number;
+    uint16_t* result_ptr = (uint16_t*)result;
 
     precision *= 2;
     while (precision--) {
         digit x = (digit)*left_number_ptr - right_number - borrow;
         left_number_ptr++;
-        *result_ptr++ = (unsigned short)x;
+        *result_ptr++ = (uint16_t)x;
         borrow = (((1L << 16) & x) != 0L);
 
         right_number = 0;
@@ -1204,20 +1204,20 @@ int XMP_Unsigned_Mult(digit* prod, const digit* multiplicand, const digit* multi
  *=============================================================================================*/
 int XMP_Unsigned_Mult_Int(digit* prod, const digit* multiplicand, short multiplier, int precision)
 {
-    const unsigned short* m2 = (const unsigned short*)multiplicand;
-    unsigned short* pr = (unsigned short*)prod;
-    unsigned long carry = 0;
+    const uint16_t* m2 = (const uint16_t*)multiplicand;
+    uint16_t* pr = (uint16_t*)prod;
+    uint32_t carry = 0;
     for (int i = 0; i < precision * 2; ++i) {
-        unsigned long p = (((unsigned long)multiplier) * *m2) + carry;
+        uint32_t p = (((uint32_t)multiplier) * *m2) + carry;
         ;
-        *pr = (unsigned short)p;
+        *pr = (uint16_t)p;
         carry = p >> 16;
         m2++;
         pr++;
     }
 
     /* Add carry to the next higher word of product / dividend */
-    //	*pr += (unsigned short)carry;
+    //	*pr += (uint16_t)carry;
     return (0);
 }
 
@@ -1343,12 +1343,12 @@ int XMP_Signed_Mult(digit* prod, const digit* multiplicand, const digit* multipl
  * HISTORY:                                                                                    *
  *   07/02/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-unsigned short XMP_Unsigned_Div_Int(digit* quotient, digit const* dividend, unsigned short divisor, int precision)
+uint16_t XMP_Unsigned_Div_Int(digit* quotient, digit const* dividend, uint16_t divisor, int precision)
 {
     if (!divisor)
         return 0; /* zero divisor means divide error */
 
-    unsigned short remainder = 0;
+    uint16_t remainder = 0;
 
     XMP_Init(quotient, 0, precision);
 
@@ -1647,7 +1647,7 @@ void XMP_Decode_ASCII(char const* str, digit* mpn, int precision)
     if (i == 0)
         return;
 
-    unsigned short radix;          /* base 2-16 */
+    uint16_t radix;                /* base 2-16 */
     switch (toupper(str[i - 1])) { /* classify radix select suffix character */
     case '.':
         radix = 10;
@@ -1748,18 +1748,18 @@ void XMP_Decode_ASCII(char const* str, digit* mpn, int precision)
  * HISTORY:                                                                                    *
  *   07/02/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-void XMP_Hybrid_Mul(unsigned short* prod, unsigned short* multiplicand, unsigned short multiplier, int precision)
+void XMP_Hybrid_Mul(uint16_t* prod, uint16_t* multiplicand, uint16_t multiplier, int precision)
 {
-    unsigned long carry = 0;
+    uint32_t carry = 0;
     for (int i = 0; i < precision; ++i) {
-        unsigned long p = (unsigned long)multiplier * *multiplicand++;
+        uint32_t p = (uint32_t)multiplier * *multiplicand++;
         p += *prod + carry;
-        *prod++ = (unsigned short)p;
+        *prod++ = (uint16_t)p;
         carry = p >> 16;
     }
 
     /* Add carry to the next higher word of product / dividend */
-    *prod += (unsigned short)carry;
+    *prod += (uint16_t)carry;
 }
 
 /***********************************************************************************************
@@ -1791,18 +1791,18 @@ void XMP_Double_Mul(digit* prod, const digit* multiplicand, const digit* multipl
     */
     XMP_Init(prod, 0, precision * 2);
 
-    const unsigned short* multiplier_ptr = (const unsigned short*)multiplier;
-    unsigned short* product_ptr = (unsigned short*)prod;
+    const uint16_t* multiplier_ptr = (const uint16_t*)multiplier;
+    uint16_t* product_ptr = (uint16_t*)prod;
 
     // Multiply multiplicand by each word in multiplier, accumulating prod.
     for (int i = 0; i < precision * 2; ++i) {
-        XMP_Hybrid_Mul(product_ptr++, (unsigned short*)multiplicand, *multiplier_ptr++, precision * 2);
+        XMP_Hybrid_Mul(product_ptr++, (uint16_t*)multiplicand, *multiplier_ptr++, precision * 2);
     }
 }
 
-static int _modulus_shift;                    // number of bits for recip scaling
-static unsigned short _reciprical_high_digit; // MSdigit of scaled recip
-static unsigned short _reciprical_low_digit;  // LSdigit of scaled recip
+static int _modulus_shift;              // number of bits for recip scaling
+static uint16_t _reciprical_high_digit; // MSdigit of scaled recip
+static uint16_t _reciprical_low_digit;  // LSdigit of scaled recip
 
 static int _modulus_sub_precision;                 //	length of modulus in MULTUNITs
 static int _modulus_bit_count;                     //	number of modulus significant bits
@@ -1866,7 +1866,7 @@ int XMP_Prepare_Modulus(const digit* n_modulus, int precision)
         XMP_Shift_Right_Bits(_mod_quotient, 1, 2);
         _modulus_shift--; /* now  0 <= _modulus_shift <= 16 */
     }
-    unsigned short* mpm = (unsigned short*)_mod_quotient;
+    uint16_t* mpm = (uint16_t*)_mod_quotient;
     _reciprical_low_digit = *mpm++;
     _reciprical_high_digit = *mpm;
 
@@ -1928,9 +1928,9 @@ int XMP_Mod_Mult(digit* prod, const digit* multiplicand, const digit* multiplier
         int nqd = dmi + 1 - _modulus_sub_precision; // number of quotient digits remaining to be generated
 
         /* Set msb, lsb, and normal ptrs of dividend */
-        unsigned short* dmph =
-            ((unsigned short*)_double_staging_number) + dmi + 1; // points to one higher than precision would indicate
-        unsigned short* dmpl = dmph - _modulus_sub_precision;
+        uint16_t* dmph =
+            ((uint16_t*)_double_staging_number) + dmi + 1; // points to one higher than precision would indicate
+        uint16_t* dmpl = dmph - _modulus_sub_precision;
 
         /*
         ** Divide loop.
@@ -1945,16 +1945,16 @@ int XMP_Mod_Mult(digit* prod, const digit* multiplicand, const digit* multiplier
             --dmph;
             --dmpl;
 
-            unsigned short q = mp_quo_digit(dmph); // trial quotient digit
+            uint16_t q = mp_quo_digit(dmph); // trial quotient digit
             if (q > 0) {
-                XMP_Hybrid_Mul(dmpl, (unsigned short*)_scratch_modulus, q, precision * 2);
+                XMP_Hybrid_Mul(dmpl, (uint16_t*)_scratch_modulus, q, precision * 2);
 
                 /* Perform correction if q too large.
                 **  This rarely occurs.
                 */
                 if (!(*dmph & SEMI_UPPER_MOST_BIT)) {
-                    unsigned short* dmp = dmpl;
-                    if (XMP_Sub((unsigned long*)dmp, (unsigned long*)dmp, _scratch_modulus, false, precision)) {
+                    uint16_t* dmp = dmpl;
+                    if (XMP_Sub((uint32_t*)dmp, (uint32_t*)dmp, _scratch_modulus, false, precision)) {
                         (*dmph)--;
                     }
                 }
@@ -2004,7 +2004,7 @@ void XMP_Mod_Mult_Clear(int precision)
 ** "digit" (MULTUNIT-sized digit) by multiplying the three most
 ** significant MULTUNITs of the dividend by the two most significant
 ** MULTUNITs of the reciprocal of the modulus.  Note that this function
-** requires that 16 * 2 <= sizeof(unsigned long).
+** requires that 16 * 2 <= sizeof(uint32_t).
 **
 ** An important part of this technique is that the quotient never be
 ** too small, although it may occasionally be too large.  This was
@@ -2024,30 +2024,30 @@ void XMP_Mod_Mult_Clear(int precision)
 **      three MULTUNITs at dividend by the upper two MULTUNITs of the
 **      modulus.
 */
-unsigned short mp_quo_digit(unsigned short* dividend)
+uint16_t mp_quo_digit(uint16_t* dividend)
 {
-    unsigned long q, q0, q1, q2;
+    uint32_t q, q0, q1, q2;
 
     /*
      * Compute the least significant product group.
      * The last terms of q1 and q2 perform upward rounding, which is
      * needed to guarantee that the result not be too small.
      */
-    q1 = (dividend[-2] ^ SEMI_MASK) * (unsigned long)_reciprical_high_digit + _reciprical_high_digit;
-    q2 = (dividend[-1] ^ SEMI_MASK) * (unsigned long)_reciprical_low_digit + (1L << 16);
+    q1 = (dividend[-2] ^ SEMI_MASK) * (uint32_t)_reciprical_high_digit + _reciprical_high_digit;
+    q2 = (dividend[-1] ^ SEMI_MASK) * (uint32_t)_reciprical_low_digit + (1L << 16);
     q0 = (q1 >> 1) + (q2 >> 1) + 1;
 
     /*      Compute the middle significant product group.   */
-    q1 = (dividend[-1] ^ SEMI_MASK) * (unsigned long)_reciprical_high_digit;
-    q2 = (dividend[0] ^ SEMI_MASK) * (unsigned long)_reciprical_low_digit;
+    q1 = (dividend[-1] ^ SEMI_MASK) * (uint32_t)_reciprical_high_digit;
+    q2 = (dividend[0] ^ SEMI_MASK) * (uint32_t)_reciprical_low_digit;
     q = (q0 >> 16) + (q1 >> 1) + (q2 >> 1) + 1;
 
     /*      Compute the most significant term and add in the others */
-    q = (q >> (16 - 2)) + (((dividend[0] ^ SEMI_MASK) * (unsigned long)_reciprical_high_digit) << 1);
+    q = (q >> (16 - 2)) + (((dividend[0] ^ SEMI_MASK) * (uint32_t)_reciprical_high_digit) << 1);
     q >>= _modulus_shift;
 
     /*      Prevent overflow and then wipe out the intermediate results. */
-    return (unsigned short)std::min(q, (unsigned long)(1L << 16) - 1);
+    return (uint16_t)std::min(q, (uint32_t)(1L << 16) - 1);
 }
 
 /*
@@ -2140,9 +2140,9 @@ int xmp_exponent_mod(digit* expout, const digit* expin, const digit* exponent_pt
 
 int pfunc(const void* pkey, const void* base)
 {
-    if (*(unsigned short*)pkey < *(unsigned short*)base)
+    if (*(uint16_t*)pkey < *(uint16_t*)base)
         return (-1);
-    if (*(unsigned short*)pkey > *(unsigned short*)base)
+    if (*(uint16_t*)pkey > *(uint16_t*)base)
         return (1);
     return (0);
 }
@@ -2175,8 +2175,8 @@ bool XMP_Is_Small_Prime(const digit* candidate, int precision)
     if (*candidate > primeTable[ARRAY_SIZE(primeTable) - 1])
         return false;
 
-    unsigned long* ptr =
-        (unsigned long*)bsearch(&candidate, &primeTable[0], ARRAY_SIZE(primeTable), sizeof(primeTable[0]), pfunc);
+    uint32_t* ptr =
+        (uint32_t*)bsearch(&candidate, &primeTable[0], ARRAY_SIZE(primeTable), sizeof(primeTable[0]), pfunc);
     return (ptr != NULL);
 }
 
@@ -2458,7 +2458,7 @@ bool XMP_Is_Prime(digit const* prime, int precision)
 /*
 **	Complete list of all prime numbers that are less than 32719 (inclusive).
 */
-unsigned short primeTable[3511] = {
+uint16_t primeTable[3511] = {
     0x0002, 0x0003, 0x0005, 0x0007, 0x000B, 0x000D, 0x0011, 0x0013, 0x0017, 0x001D, 0x001F, 0x0025, 0x0029, 0x002B,
     0x002F, 0x0035, 0x003B, 0x003D, 0x0043, 0x0047, 0x0049, 0x004F, 0x0053, 0x0059, 0x0061, 0x0065, 0x0067, 0x006B,
     0x006D, 0x0071, 0x007F, 0x0083, 0x0089, 0x008B, 0x0095, 0x0097, 0x009D, 0x00A3, 0x00A7, 0x00AD, 0x00B3, 0x00B5,
